This commis will help to determine the cause of the issue GH-40.
The error has been present since the beginning and I thought I did something stupid (still possible and the case) but given the "random" nature of the issue and the number of queue involved it is not easy to identify the cause.

Having insight of the content of the payload received from the queue will  give help us to determine what we are receiving in those situations that are causing the app to fail and  from there we can investigate.